Firebase initialization in main
