+++ src/main/java/use_case/signup/SignupInteractor.java
@@ -1,5 +1,7 @@
 package use_case.signup;
 
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.UserRecord;
 import entity.User;
 import entity.UserFactory;
 
@@ -31,8 +33,21 @@ public class SignupInteractor implements SignupInputBoundary {
             final User user = userFactory.create(signupInputData.getUsername(), signupInputData.getPassword());
             userDataAccessObject.save(user);
 
+            signUpFirebaseUser(signupInputData.getUsername(), signupInputData.getPassword());
             final SignupOutputData signupOutputData = new SignupOutputData(user.getName(), false);
             userPresenter.prepareSuccessView(signupOutputData);
+        }
+    }
+    public void signUpFirebaseUser(String email, String password) {
+        try {
+            UserRecord.CreateRequest request = new UserRecord.CreateRequest()
+                    .setEmail(email)
+                    .setPassword(password);
+            UserRecord userRecord = FirebaseAuth.getInstance().createUser(request);
+            System.out.println("Successfully created user: " + userRecord.getUid());
+        }
+        catch (Exception e) {
+            e.printStackTrace();
         }
     }
